Windows loves being different

+++ src/extension/IOController_Windows.hpp
@@ -105,8 +105,8 @@ namespace extension {
                     // Check how many bytes are available to read, if it's 0 and we have a read event the
                     // descriptor is sending EOF and we should fire a CLOSE event too and stop watching
                     if ((wsae.lNetworkEvents & IO::READ) != 0) {
-                        int bytes_available = 0;
-                        bool valid          = ::ioctlsocket(r->second.fd, FIONREAD, &bytes_available) == 0;
+                        u_long bytes_available = 0;
+                        bool valid             = ::ioctlsocket(r->second.fd, FIONREAD, &bytes_available) == 0;
                         if (valid && bytes_available == 0) {
                             wsae.lNetworkEvents |= IO::CLOSE;
                         }
